move all commands to command dir
