--- webdriver/src/commands.ts
@@ -1,91 +1,0 @@
-import type { IncomingMessage, ServerResponse } from "http";
-
-// BELOW ARE ACTIONS FROM TESTS THEMSELVES
-import type { WebdriverParams } from "./config.js";
-import type { EventBusInterface } from "./eventbus.js";
-
-import * as fs from "fs";
-import * as path from "path";
-
-let headers = new Headers([["Content-Type", "application/json"]]);
-
-// GO
-export async function go(
-	params: WebdriverParams,
-	signal: AbortSignal,
-	sessionId: string,
-	hostAndPort: URL,
-	targetPath: string,
-): Promise<void> {
-	let { url } = params;
-
-	let pingUrl = new URL(targetPath, hostAndPort);
-	let getCookie = await fetch(new URL(`/session/${sessionId}/url`, url), {
-		method: "POST",
-		headers,
-		body: JSON.stringify({ url: pingUrl }),
-		signal,
-	});
-
-	if (200 !== getCookie.status) {
-		let cause = await getCookie.json();
-		throw new Error("go-to request failed", { cause });
-	}
-}
-
-// ADD COOKIE
-export async function addCookie(
-	params: WebdriverParams,
-	signal: AbortSignal,
-	sessionId: string,
-): Promise<void> {
-	let { url, jackrabbitId } = params;
-
-	let cookieReq = await fetch(new URL(`/session/${sessionId}/cookie`, url), {
-		method: "POST",
-		headers,
-		body: JSON.stringify({
-			cookie: {
-				name: "jackrabbit",
-				value: jackrabbitId,
-				path: "/",
-				httpOnly: true,
-			},
-		}),
-		signal,
-	});
-
-	if (200 !== cookieReq.status) {
-		let cause = await cookieReq.json();
-		throw new Error("set-cookie request failed", { cause });
-	}
-}
-
-// FIND ELEMENT
-
-function getJsonFromRequestBody(req: IncomingMessage): Promise<any> {
-	return new Promise(function (resolve, reject) {
-		let data: Uint8Array[] = [];
-		req.addListener("data", function (chunk) {
-			data.push(chunk);
-		});
-		req.addListener("end", function () {
-			let actionStr = Buffer.concat(data).toString();
-			let action = JSON.parse(actionStr);
-
-			resolve(action);
-		});
-		req.addListener("error", function (err: Error) {
-			reject(err);
-		});
-	});
-}
-
-export async function findElementFromElement() {}
-export async function findElementsFromElements() {}
-
-// export async function findShadowRoot() {}
-export async function findElementFromShadowRoot() {}
-export async function findElementsFromShadowRoot() {}
-
-// LOG
