+++ webdriver/src/commands/add_cookie.ts
@@ -1,0 +1,34 @@
+import type { IncomingMessage, ServerResponse } from "http";
+
+// BELOW ARE ACTIONS FROM TESTS THEMSELVES
+import type { WebdriverParams } from "../config.js";
+
+import { jsonHeaders } from "./flyweight.js";
+
+// ADD COOKIE
+export async function addCookie(
+	params: WebdriverParams,
+	signal: AbortSignal,
+	sessionId: string,
+): Promise<void> {
+	let { url, jackrabbitId } = params;
+
+	let cookieReq = await fetch(new URL(`/session/${sessionId}/cookie`, url), {
+		method: "POST",
+		headers: jsonHeaders,
+		body: JSON.stringify({
+			cookie: {
+				name: "jackrabbit",
+				value: jackrabbitId,
+				path: "/",
+				httpOnly: true,
+			},
+		}),
+		signal,
+	});
+
+	if (200 !== cookieReq.status) {
+		let cause = await cookieReq.json();
+		throw new Error("set-cookie request failed", { cause });
+	}
+}
