+++ webdriver/src/commands/mod.ts
@@ -6,3 +6,12 @@ export * from "./get_element_shadow_root.js";
 export * from "./log.js";
 export * from "./sessions.js";
 export * from "./take_element_screenshot.js";
+export * from "./navigate_to.js";
+export * from "./add_cookie.js";
+
+export async function findElementFromElement() {}
+export async function findElementsFromElements() {}
+
+// export async function findShadowRoot() {}
+export async function findElementFromShadowRoot() {}
+export async function findElementsFromShadowRoot() {}
