+++ webdriver/src/commands/navigate_to.ts
@@ -1,0 +1,29 @@
+import type { IncomingMessage, ServerResponse } from "http";
+
+// BELOW ARE ACTIONS FROM TESTS THEMSELVES
+import type { WebdriverParams } from "../config.js";
+
+import { jsonHeaders } from "./flyweight.js";
+
+export async function navigateTo(
+	params: WebdriverParams,
+	signal: AbortSignal,
+	sessionId: string,
+	hostAndPort: URL,
+	targetPath: string,
+): Promise<void> {
+	let { url } = params;
+
+	let pingUrl = new URL(targetPath, hostAndPort);
+	let getCookie = await fetch(new URL(`/session/${sessionId}/url`, url), {
+		method: "POST",
+		headers: jsonHeaders,
+		body: JSON.stringify({ url: pingUrl }),
+		signal,
+	});
+
+	if (200 !== getCookie.status) {
+		let cause = await getCookie.json();
+		throw new Error("go-to request failed", { cause });
+	}
+}
